Move `RuntimePermissionResult` enum to the `:simple-platform` module
This is in preparation for a future commit where we will move the
`Analytics` class to the  `:simple-platform` module.
