+++ app/src/main/java/org/simple/clinic/contactpatient/ContactPatientEvent.kt
@@ -6,7 +6,7 @@ import org.simple.clinic.patient.PatientProfile
 import org.simple.clinic.util.None
 import org.simple.clinic.util.Optional
 import org.simple.clinic.util.RequiresPermission
-import org.simple.clinic.util.RuntimePermissionResult
+import org.simple.clinic.platform.util.RuntimePermissionResult
 import org.simple.clinic.widgets.UiEvent
 import org.threeten.bp.LocalDate
