+++ app/src/main/java/org/simple/clinic/facility/change/FacilityChangeActivityController.kt
@@ -19,8 +19,8 @@ import org.simple.clinic.location.LocationUpdate.Available
 import org.simple.clinic.location.LocationUpdate.Unavailable
 import org.simple.clinic.user.UserSession
 import org.simple.clinic.util.ElapsedRealtimeClock
-import org.simple.clinic.util.RuntimePermissionResult.DENIED
-import org.simple.clinic.util.RuntimePermissionResult.GRANTED
+import org.simple.clinic.platform.util.RuntimePermissionResult.DENIED
+import org.simple.clinic.platform.util.RuntimePermissionResult.GRANTED
 import org.simple.clinic.util.filterAndUnwrapJust
 import org.simple.clinic.util.timer
 import org.simple.clinic.widgets.ScreenCreated
