+++ app/src/main/java/org/simple/clinic/home/patients/PatientsScreenEvents.kt
@@ -4,7 +4,7 @@ import android.Manifest
 import org.simple.clinic.util.None
 import org.simple.clinic.util.Optional
 import org.simple.clinic.util.RequiresPermission
-import org.simple.clinic.util.RuntimePermissionResult
+import org.simple.clinic.platform.util.RuntimePermissionResult
 import org.simple.clinic.widgets.UiEvent
 
 object NewPatientClicked : UiEvent {
