+++ app/src/main/java/org/simple/clinic/registration/location/RegistrationLocationPermissionScreenEvents.kt
@@ -4,7 +4,7 @@ import org.simple.clinic.location.LOCATION_PERMISSION
 import org.simple.clinic.util.None
 import org.simple.clinic.util.Optional
 import org.simple.clinic.util.RequiresPermission
-import org.simple.clinic.util.RuntimePermissionResult
+import org.simple.clinic.platform.util.RuntimePermissionResult
 import org.simple.clinic.widgets.UiEvent
 
 data class RequestLocationPermission(
