+++ app/src/main/java/org/simple/clinic/scheduleappointment/facilityselection/FacilitySelectionActivityEvents.kt
@@ -2,7 +2,7 @@ package org.simple.clinic.scheduleappointment.facilityselection
 
 import org.simple.clinic.facility.Facility
 import org.simple.clinic.location.LocationUpdate
-import org.simple.clinic.util.RuntimePermissionResult
+import org.simple.clinic.platform.util.RuntimePermissionResult
 import org.simple.clinic.widgets.UiEvent
 
 data class FacilitySelected(val facility: Facility) : UiEvent {
