+++ app/src/test/java/org/simple/clinic/analytics/AnalyticsTest.kt
@@ -5,8 +5,8 @@ import org.junit.After
 import org.junit.Test
 import org.simple.clinic.analytics.MockAnalyticsReporter.Event
 import org.simple.clinic.platform.analytics.AnalyticsUser
-import org.simple.clinic.util.RuntimePermissionResult.DENIED
-import org.simple.clinic.util.RuntimePermissionResult.GRANTED
+import org.simple.clinic.platform.util.RuntimePermissionResult.DENIED
+import org.simple.clinic.platform.util.RuntimePermissionResult.GRANTED
 import org.threeten.bp.Duration
 import org.threeten.bp.Instant
 import java.util.UUID
