+++ app/src/test/java/org/simple/clinic/contactpatient/ContactPatientUpdateTest.kt
@@ -27,8 +27,8 @@ import org.simple.clinic.overdue.TimeToAppointment.Days
 import org.simple.clinic.overdue.TimeToAppointment.Weeks
 import org.simple.clinic.phone.PhoneNumberMaskerConfig
 import org.simple.clinic.util.Just
-import org.simple.clinic.util.RuntimePermissionResult.DENIED
-import org.simple.clinic.util.RuntimePermissionResult.GRANTED
+import org.simple.clinic.platform.util.RuntimePermissionResult.DENIED
+import org.simple.clinic.platform.util.RuntimePermissionResult.GRANTED
 import org.simple.clinic.util.TestUserClock
 import org.threeten.bp.LocalDate
 import org.threeten.bp.Period
