+++ app/src/test/java/org/simple/clinic/util/RequestPermissionsTest.kt
@@ -9,13 +9,14 @@ import com.nhaarman.mockitokotlin2.whenever
 import io.reactivex.subjects.PublishSubject
 import org.junit.After
 import org.junit.Test
+import org.simple.clinic.platform.util.RuntimePermissionResult
 import org.simple.clinic.router.screen.ActivityPermissionResult
 import org.simple.clinic.util.RequestPermissionsTest.Event.FirstEvent
 import org.simple.clinic.util.RequestPermissionsTest.Event.FourthEvent
 import org.simple.clinic.util.RequestPermissionsTest.Event.SecondEvent
 import org.simple.clinic.util.RequestPermissionsTest.Event.ThirdEvent
-import org.simple.clinic.util.RuntimePermissionResult.DENIED
-import org.simple.clinic.util.RuntimePermissionResult.GRANTED
+import org.simple.clinic.platform.util.RuntimePermissionResult.DENIED
+import org.simple.clinic.platform.util.RuntimePermissionResult.GRANTED
 
 class RequestPermissionsTest {
