+++ simple-platform/src/main/java/org/simple/clinic/platform/util/RuntimePermissionResult.kt
@@ -1,0 +1,6 @@
+package org.simple.clinic.platform.util
+
+enum class RuntimePermissionResult {
+  GRANTED,
+  DENIED
+}
